--- packages/renderer/src/components/Editor/PublishModal/index.ts
@@ -1,2 +1,0 @@
-import { PublishModal } from './component';
-export { PublishModal };
