+++ packages/renderer/src/components/Modals/DeleteProject/index.ts
@@ -1,0 +1,2 @@
+import { DeleteProject } from './component';
+export { DeleteProject };
