+++ packages/renderer/src/components/Modals/DeleteProject/types.ts
@@ -1,0 +1,8 @@
+import { type Project } from '/shared/types/projects';
+
+export type Props = {
+  open: boolean;
+  project: Project;
+  onClose: () => void;
+  onSubmit: (project: Project) => void;
+};
