+++ packages/renderer/src/components/Modals/PublishProject/index.ts
@@ -1,0 +1,3 @@
+import { PublishProject } from './component';
+import type { StepValue } from './types';
+export { PublishProject, StepValue };
